+++ src/output/Phrase.ts
@@ -37,6 +37,8 @@ import {
     layoutToCSS,
     type WritingLayoutSymbol,
 } from '@locale/Scripts';
+import { toAura } from './Aura';
+import type Aura from './Aura';
 
 export function createPhraseType(locales: Locales) {
     return toStructure(`
@@ -92,6 +94,7 @@ export function createPhraseType(locales: Locales) {
             (locale) => locale.output.Phrase.direction,
         )}•'${HorizontalLayout}'|'${VerticalRightLeftLayout}'|'${VerticalLeftRightLayout}': '${HorizontalLayout}'
         ${getBind(locales, (locale) => locale.output.Phrase.matter)}•Matter|ø: ø
+        ${getBind(locales, (locale) => locale.output.Phrase.aura)}•ø|🔮: ø
     )`);
 }
 
@@ -112,6 +115,7 @@ export default class Phrase extends Output {
     readonly alignment: string | undefined;
     readonly direction: WritingLayoutSymbol;
     readonly matter: Matter | undefined;
+    readonly aura: Aura | undefined;
 
     private _metrics: Metrics | undefined = undefined;
 
@@ -137,6 +141,7 @@ export default class Phrase extends Output {
         alignment: string | undefined,
         direction: WritingLayoutSymbol,
         matter: Matter | undefined,
+        aura: Aura | undefined,
     ) {
         super(
             value,
@@ -160,6 +165,7 @@ export default class Phrase extends Output {
         this.alignment = alignment;
         this.direction = direction;
         this.matter = matter;
+        this.aura = aura;
 
         // Make sure this font is loaded. This is a little late -- we could do some static analysis
         // and try to determine this in advance -- but anything can compute a font name. Maybe an optimization later.
@@ -415,6 +421,7 @@ export function toPhrase(
     const alignment = toText(getOutputInput(value, 21));
     const direction = toText(getOutputInput(value, 22));
     const matter = toMatter(getOutputInput(value, 23));
+    const shadow = toAura(project, getOutputInput(value, 24));
 
     return texts !== undefined &&
         duration !== undefined &&
@@ -442,6 +449,7 @@ export function toPhrase(
               alignment?.text,
               direction.text as WritingLayoutSymbol,
               matter,
+              shadow,
           )
         : undefined;
 }